requirements.txt
Django==4.2.14
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
djangorestframework~=3.15.2
django-debug-toolbar~=4.4.6
pillow==10.4.0
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
factory-boy==3.3.0
